Forward ansible_ssh_port to SSH ProxyCommand
This inventory template sets up a proxy command to SSH into the guest
via the host. It forwards the ansible_user and ansible_host but it
doesn't forward the ansible_ssh_port, which means SSHing to the VM fails
to connect if the host is using a nonstandard SSH port.