Make network non-optional
This is kinda annoying because it's specific to my current hardware that
en1 is plugged in an en0 is not. If I make both devices required then
the boot takes 2 extra minutes while systemd-networkd waits for an IP
address on both devices.

However if I make both optional then network-online.target seems to
come up too fast and sometimes the SSH tunneling doesn't work.