Try removing unused ethernet
My SSH tunnel hackery is unreliable, sometimes when the system boots the
tunnel doesn't work and I have to restart it.

My theory here was that this was because the network-online.target was
getting hit early because of the optional network interface. But, that
doesn't seem to be the issue.