Document the gunslinger lifestyle
"Every child knows that play is nobler than work. He knows too that the
worth or merit of a game is not inherent in the game itself but rather
in the value of that which is put at hazard. Games of chance require a
wager to have meaning at all. Games of sport involve the skill and
strength of the opponents and the humiliation of defeat and the pride of
victory are in themselves sufficient stake because they inhere in the
worth of the principals and define them. But trial of chance or trial of
worth all games aspire to the condition of war for here that which is
wagered swallows up game, player, all."